requirements.txt
requests
google-generativeai
selenium
webdriver-manager
python-dotenv
schedule
peewee
fake-useragent
loguru
